app/requirements.txt
flask
pywebview
pandas
pyinstaller
waitress